requirements.txt
pydantic==2.5.0
openai==1.3.0
pyyaml==6.0.1
python-dotenv==1.0.0